Development

+++ extension/cmd.js
@@ -8,8 +8,7 @@ const session = require('./session.js');
 const dock = require('./dockerode.js');
 
 function info(sessionId) {
-    const sessionData = session.map.get(sessionId);
-    const { sessionIp, containerId, containerName, isDebug, isLoggedIn, isAdmin, screenWidth, screenHeight, timestamp } = sessionData ?? '';
+    const { sessionIp, containerId, containerName, isDebug, isLoggedIn, isAdmin, screenWidth, screenHeight, timestamp } = session.read(sessionId, ['sessionIp', 'containerId', 'containerName', 'isDebug', 'isLoggedIn', 'isAdmin', 'screenWidth', 'screenHeight', 'timestamp']);
     return  `console.warn("` +
             `Session ID: ${sessionId}\\n` +
             `Session IP: ${sessionIp}\\n` +
@@ -58,8 +57,7 @@ function theme(socket) { // Invert everything
     socket.emit('eval', themeScript);
 }
 
-function reset(socket, sessionId) { // Reset the client page
-    session.update(sessionId, { containerName: null });
+function reset(socket) { // Reset the client page
     const resetScript = fs.readFileSync('./stream/reset.js', 'utf8');
     socket.emit('eval', resetScript);
 }
@@ -77,9 +75,9 @@ async function terminal(socket, sessionId, local = false) { // Create terminal f
     // Use the stored screen size information to calculate cols and rows
     const cols = Math.floor(screenWidth / 8); // Character width of 8px
     const rows = Math.floor(screenHeight / 16); // Character height of 16px
-
+    
     process.env['TERM'] = 'xterm-256color'; // Support 256-bit color depth
-    let shell;
+    let shell = null;
     if (local) { // Create local terminal
         const isWin32 = !!(os.platform() === 'win32');
         shell = pty.spawn(isWin32 ? 'cmd.exe' : 'bash', [], {
@@ -91,48 +89,53 @@ async function terminal(socket, sessionId, local = false) { // Create terminal f
         });
         shell.write(isWin32 ? 'echo Hello!\r\n' : 'clear && cat /etc/motd\n');
     }
-    if (!containerId) { // Spawn new container if containerId is missing
+    if (!local && !containerId) { // Spawn new container if containerId is missing
         containerId = await dock.containerCreate(); // Create new container
         containerName = await dock.containerGetName(containerId); // Retrieve generated container name
+    }
+
+    // Proceed if container is running
+    const isContainerRunning = await dock.containerRunning(containerId);
+    if (isContainerRunning) {
         session.update(sessionId, { containerId: containerId, containerName: containerName });
+
+        if (!local && containerName) { // Attach container console
+            shell = pty.spawn('docker', ['attach', `${containerName}`], {
+                name: 'xterm-color',
+                cols: cols,
+                rows: rows,
+                cwd: process.env.HOME,
+                env: process.env
+            });
+        } 
+
+        if (shell) {
+            // Pipe the output of the pseudo-terminal to the socket reaching the client
+            shell.onData(data => {
+                socket.emit('terminalOutput', data);
+            });
+
+            // Handle input from the client and write it to the pseudo-terminal
+            socket.on('terminalInput', input => {
+                shell.write(input);
+            });
+
+            // Listen for the exit event
+            shell.on('exit', () => {
+                console.info(`Terminal "${containerName || 'local'}" closed from ${sessionId}`);
+                shell.kill();
+                reset(socket, sessionId);
+            });
+            
+            // Return the JavaScript code to create the terminal on client-side
+            const terminalScript = fs.readFileSync('./stream/terminal.js', 'utf8');
+            socket.emit('eval', terminalScript);
+            console.info(`Terminal "${containerName || 'local'}" attached to ${sessionId}`);
+        }
+    } else { // Throw generic error
+        await dock.containerRemove(containerId);
+        socket.emit('eval', `console.warn("Couldn't connect to terminal");`);
     }
-
-    let {sesIp} = session.read(sessionId, ['sessionIp']);
-    console.warn(sesIp  + " vs " + sessionId);
-    
-    if (containerName) { // Attach container console
-        shell = pty.spawn('docker', ['attach', `${containerName}`], {
-            name: 'xterm-color',
-            cols: cols,
-            rows: rows,
-            cwd: process.env.HOME,
-            env: process.env
-        });
-    }
-
-    if (shell) {
-        // Pipe the output of the pseudo-terminal to the socket
-        shell.onData(data => {
-            // Emit the data to the client
-            socket.emit('terminalOutput', data);
-        });
-
-        // Handle input from the client and write it to the pseudo-terminal
-        socket.on('terminalInput', input => {
-            shell.write(input);
-        });
-
-        // Listen for the exit event
-        shell.on('exit', () => {
-            console.info(`Terminal "${containerName || 'local'}" closed from ${sessionId}`);
-            shell.kill();
-            reset(socket, sessionId);
-        });
-        // Return the JavaScript code to create the terminal on client-side
-        const terminalScript = fs.readFileSync('./stream/terminal.js', 'utf8');
-        socket.emit('eval', terminalScript);
-        console.info(`Terminal "${containerName || 'local'}" attached to ${sessionId}`);
-    } else socket.emit('eval', `console.warn("Couldn't connect to terminal");`);
 }
 
 module.exports = {
